requirements.txt
python-telegram-bot>=21.4
pytz>=2024.1
aiohttp>=3.9
python-dateutil>=2.9
